+++ heksher/api/v1/settings.py
@@ -69,7 +69,7 @@ async def declare_setting(input: DeclareSettingInput, app: HeksherApp = applicat
         await app.db_logic.add_setting(new_setting)
         return DeclareSettingOutput(created=True, changed=[], incomplete={})
 
-    to_change = {'last_touch_time': datetime.now()}
+    to_change = {'last_touch_time': datetime.utcnow()}
     changed = []
     incomplete = {}